requirements.txt
langchain
langgraph
langchain-groq
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic